refactored to launch a sessions view screen as the main view
